% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_file.R
\name{batch_file}
\alias{batch_file}
\title{Create a batch file for one or more CSM X files}
\usage{
batch_file(xl, xfiles, outdir, btype, bname = "R_DSSBatch", RP = 1,
  SQ = 0, OP = 0, CO = 0)
}
\arguments{
\item{xfiles}{Vector of x file names}

\item{outdir}{Output directory name}

\item{btype}{What sort of batch file, e.g. "MAIZE"}

\item{bname}{Optional batch file name--.v45 suffix is automatically appended}

\item{RP}{default 1. A CSM parameter}

\item{SQ}{default 0. For sequential runs}

\item{OP}{default 0}

\item{CO}{default 0}
}
\value{
.v45 batch file in selected directory
}
\description{
Create a batch file for one or more CSM X files
}
\note{
Batch files are hard-coded to spatial now.  Needs to change
}

