% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{bdens}
\alias{bdens}
\title{Bulk density accounting for compaction plus gravel}
\usage{
bdens(thetas, DF = 1, gravel = 0)
}
\arguments{
\item{thetas}{Saturation water content (without compaction)}

\item{DF}{Density factor between 0.9 and 1.3, normal (default) at 1}

\item{gravel}{Gravel percent by weight}
}
\description{
Bulk density accounting for compaction plus gravel
}
\keyword{internal}
