% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colinsert}
\alias{colinsert}
\title{Move a single column within a data.table}
\usage{
colinsert(colname, newpos, DT)
}
\arguments{
\item{colname}{Name or index of column to move}

\item{newpos}{Column name to left of new location, or index of new location.}

\item{DT}{data.table to operate on}
}
\value{
data.table DT with column order changed by reference
}
\description{
Move a single column within a data.table
}
\details{
This is a wrapper function for data.table's setcolorder function, 
written to avoid having to change an entire vector of column names.
}
\examples{
DT <- data.table(a = 1:10, b = 1:10, c = 11:20, d = 26:35)
colinsert("c", "a", DT)
DT <- data.table(a = 1:10, b = 1:10, c = 11:20, d = 26:35)
colinsert(3, 1, DT)
DT <- data.table(a = 1:10, b = 1:10, c = 11:20, d = 26:35)
colinsert(4, 1, DT)
DT <- data.table(a = 1:10, b = 1:10, c = 11:20, d = 26:35)
colinsert(4, 3, DT)
DT <- data.table(a = 1:10, b = 1:10, c = 11:20, d = 26:35)
colinsert("d", "b", DT)
DT <- data.table(a = 1:10, b = 1:10, c = 11:20, d = 26:35)
colinsert("d", 3, DT)
}
