% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dew_point.R
\name{dew_point}
\alias{dew_point}
\title{Calculate dewpoint temperature}
\usage{
dew_point(ea = NULL, rh = NULL, tmax = NULL, tmin = NULL, tmu = NULL,
  sh = NULL, pres = NULL)
}
\arguments{
\item{ea}{Vapor pressure (mb)}

\item{rh}{Relative humidity (\%)}

\item{tmax}{Maximum temperature (deg C)}

\item{tmin}{Minimum temperature (deg C)}

\item{tmu}{Average temperature (deg C)}

\item{sh}{Specific humidity (kg/kg)}

\item{pres}{Surface pressure (mb)}
}
\value{
Dewpoint temperature (deg C)
}
\description{
Calculate dewpoint temperature
}
\details{
This function was adapted from a python function provided by the 
The Parallel System for Integrating Impact Models and Sectors (pSIMS). 
https://github.com/RDCEP/psims
}
\references{
Elliott, J., D. Kelly, J. Chryssanthacopoulos, M. Glotter, 
Kanika Jhunjhnuwala, N. Best, M. Wilde, and I. Foster, (2014). The Parallel 
System for Integrating Impact Models and Sectors (pSIMS). Environmental 
Modeling and Software. http://dx.doi.org/10.1016/j.envsoft.2014.04.008
}
