% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_utilities.R
\name{exec_csm}
\alias{exec_csm}
\title{DSSAT45 execution command}
\usage{
exec_csm(projdir, csmdir, rundir, btype = "B", bname = "R_DSSBatch.v45")
}
\arguments{
\item{projdir}{Your working directory}

\item{csmdir}{Master directory for DSSAT model}

\item{rundir}{CSM where model should execute (e.g. Maize)}

\item{btype}{Batch mode code ("B" normal batch (default), "S" spatial, etc)}

\item{bname}{Batch file name, default is R_DSSBatch.v45}
}
\value{
File_path for DSSAT
}
\description{
DSSAT45 execution command
}

