% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{field_cap}
\alias{field_cap}
\title{Calculates field capacity}
\usage{
field_cap(sand, clay, soc)
}
\arguments{
\item{sand}{Fraction of sand}

\item{clay}{Fraction of clay}

\item{soc}{soil organic matter percent}
}
\value{
Field capacity at 33 kpa
}
\description{
Calculates field capacity
}
\keyword{internal}
