% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{field_cap_df}
\alias{field_cap_df}
\title{Calculated field capacity accounting for compaction}
\usage{
field_cap_df(sand, clay, soc, DF)
}
\arguments{
\item{sand}{Fraction of sand}

\item{clay}{Fraction of clay}

\item{soc}{Soil organic matter percent}

\item{DF}{Density factor between 0.9 and 1.3, normal (default) at 1}
}
\description{
Calculated field capacity accounting for compaction
}
\keyword{internal}

