% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fp}
\alias{fp}
\title{Easier joining of file paths and file names for reading in files}
\usage{
fp(path, fname)
}
\arguments{
\item{path}{Path directory}

\item{fname}{File name}
}
\value{
Complete file string
}
\description{
Easier joining of file paths and file names for reading in files
}
\details{
Function imported from lmisc, to avoid need for importing
}
\keyword{internal}
