% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{ksat}
\alias{ksat}
\title{Saturated hydraulic conductivity, including gravel effects.}
\usage{
ksat(sand, clay, soc, DF = 1, gravel = 0)
}
\arguments{
\item{sand}{Fraction of sand}

\item{clay}{Fraction of clay}

\item{soc}{Soil organic matter percent}

\item{DF}{Density factor between 0.9 and 1.3, normal (default) at 1}

\item{gravel}{Gravel percent by weight (0 by default)}
}
\description{
Saturated hydraulic conductivity, including gravel effects.
}
\keyword{internal}

