% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csm_outfile.R
\name{read_csm_outfile}
\alias{read_csm_outfile}
\title{Read in output file from DSSAT CSM}
\usage{
read_csm_outfile(rundir, type = "summary", vars)
}
\arguments{
\item{rundir}{CSM where model should execute (e.g. Maize)}

\item{type}{Specify which CSM output file type. Default is "summary"}

\item{vars}{A vector of specific variables from the file of interest}
}
\value{
Numeric matrix containing year, day, and specified variables
}
\description{
Read in output file from DSSAT CSM
}
\details{
This is a read.table based variant of read_csm_outfile, designed for 
reading in DSSAT .OUT files, particularly Summary.OUT files, but 
theoretically including (although not yet tested on) other, e.g. ET.OUT.  
The vars variable should be provide the exact variable names of interest. 
E.g. "EOAC and EOAA" from ET.OUT.
}
\note{
This is only set up to read in the Summary.OUT file at present
}
