% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sol_hor.R
\name{read_sol_hor}
\alias{read_sol_hor}
\title{Read iprofile data from DSSAT .SOL file}
\usage{
read_sol_hor(solfile, profiles)
}
\arguments{
\item{solfile}{Name of .SOL file to read in}

\item{profiles}{Vector of profile name to read in (must be exact)}
}
\value{
List of soil parameters of interest
}
\description{
This reads in horizon depths and water-holding variables from a 
DSSAT .SOL file, for use in initializing the experimental files
}
\details{
This function is stripped-down version of read_sol_prof, intended to 
pull out just the Horizon depths, drained upper and lower limits, so that 
these data can be used to initialize starting soil moisture conditions in 
DSSAT x files. Complete profiles can be read using read_sol_prof.
}
\examples{
f <- "~/DSSAT45/SOIL/WI.SOL"  # .SOL file (WISE)
read_sol_hor(solfile = f, profiles = "WI_ACLS021") 
}
\keyword{DSSAT}
