% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sol_prof.R
\name{read_sol_prof}
\alias{read_sol_prof}
\title{Read in profile data from DSSAT .SOL file}
\usage{
read_sol_prof(solfile, profiles)
}
\arguments{
\item{solfile}{Name of .SOL file to read in}

\item{profiles}{Vector of profile name to read in (must be exact)}
}
\value{
List of soil parameters of interest
}
\description{
This reads most horizon data from an existing DSSAT soil file, 
so that the parameters can be used in experimental files or for analysis
}
\details{
This function currently pulls nearly all the horizon data from a 
DSSAT .sol profile. It is a fuller version of read_SOL_prof
}
\examples{
f <- "~/DSSAT45/SOIL/WI.SOL"  # .SOL file (WISE)
sprofs <- read_sol_prof(solfile = f, profiles = "WI_ACLS021")
sprofs[, c(1:7, 11:16), with = FALSE]
}
\keyword{DSSAT}
