% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{ro_df}
\alias{ro_df}
\title{Matric density accounting for compaction}
\usage{
ro_df(thetas, DF = 1)
}
\arguments{
\item{thetas}{Saturation water content}

\item{DF}{Density factor between 0.9 and 1.3, normal (default) at 1}
}
\value{
Matric density
}
\description{
Matric density accounting for compaction
}
\keyword{internal}

