% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_soil_utilities.R
\name{sldr}
\alias{sldr}
\title{Soil drainage rate}
\usage{
sldr(clay)
}
\arguments{
\item{clay}{Clay percentage}
}
\description{
Soil drainage rate
}
\details{
Based on Estes et al (2013), following Schulze et al for ACRU model.
Region-specific functions might be better to use
}
\keyword{internal}
