% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_soil_utilities.R
\name{slro}
\alias{slro}
\title{Calcutes the runoff curve number following Gijsman et al (2007)}
\usage{
slro(slo, depth, texture, drainage)
}
\arguments{
\item{slo}{Slope in percent}

\item{depth}{Total depth of profile (in cm)}

\item{drainage}{Soil drainage rate (from SLDR function)}

\item{textclass}{Texture class (from soil_class function)}
}
\description{
Calcutes the runoff curve number following Gijsman et al (2007)
}
\keyword{internal}

