% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_soil_utilities.R
\name{slu1}
\alias{slu1}
\title{Evaporation limit function from Du Toit, source unknown}
\usage{
slu1(clay1)
}
\arguments{
\item{clay1}{Clay percentage for the top soil horizon}
}
\description{
Evaporation limit function from Du Toit, source unknown
}
\keyword{internal}

