% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_class.R
\name{soil_class}
\alias{soil_class}
\title{Returns soil texture class, based on USDA triangle}
\usage{
soil_class(clay, sand, silt = NULL)
}
\arguments{
\item{clay}{Vector of percent clay}

\item{sand}{Vector of percent silt}

\item{silt}{Vector of percent sand}
}
\value{
USDA texture class
}
\description{
Returns soil texture class, based on USDA triangle
}
\details{
This function draws on tt_class, and adaptation of the 
TT.points.in.classes from the soiltexture package. That function can classify
borderline soiltexture cases. This function returns just the first of several
classifications made.
}
\examples{
soil_class(clay = c(05,60,15,05,25,05,25,45,65,75,13,47),
           sand = c(90,32,70,70,20,10,10,10,20,10,70,10))
}
\references{
Julien Moeys (2016). soiltexture: Functions for Soil Texture
Plot, Classification and Transformation. R package version 1.4.1. 
https://CRAN.R-project.org/package=soiltexture
}
