% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{soil_hydraulics}
\alias{soil_hydraulics}
\title{Calculates various soil hydraulic properties, following Saxton & Rawls, 2006}
\usage{
soil_hydraulics(sand, clay, soc, DF = 1, gravel = 0, digits = 4,
  PAW = TRUE)
}
\arguments{
\item{sand}{Fraction of sand}

\item{clay}{Fraction of clay}

\item{soc}{Soil organic matter percent}

\item{DF}{Density factor between 0.9 and 1.3, normal (default) at 1}

\item{gravel}{Gravel percent by weight (0 by default)}

\item{digits}{Number of significant digits (4 by default)}

\item{PAW}{Gravel and density adjusted plant available water (TRUE or FALSE)}
}
\description{
Calculates various soil hydraulic properties, following Saxton & Rawls, 2006
}
\details{
A single function producing estimates of wilting point, 
field capacity, saturated water content, bulk density, and saturdated 
hydraulic conductivity, account for soil density and gravel effects, based on 
methods described by Saxton and Rawls (2006). Internal functions for each 
variables can be also used separately, as needed. Per Saxton & Rawls (2006), 
these functions are only valid for SOC <= 8% clay <= 60%. Functions were 
checked against equations available for download with SPAW model, 
downloadable at http://hrsl.arsusda.gov/SPAW/SPAWDownload.html.
}
\examples{
soil_hydraulics(sand = 0.29, clay = 0.32, soc = 3.51, DF = 1, gravel = 0)
soil_hydraulics(sand = 0.29, clay = 0.32, soc = 3.51, DF = 0.8, gravel = 0)
soil_hydraulics(sand = 0.29, clay = 0.32, soc = 3.51, DF = 1, gravel = 0.2)
soil_hydraulics(sand = 0.29, clay = 0.32, soc = 3.51, DF = 0.8, gravel = 0.2)
}
\references{
Saxton, K.E. & Rawls, W.J. (2006) Soil water characteristic estimates by 
texture and organic matter for hydrologic solutions. Soil Sci Soc Am J, 70, 
1569–1578.
}
