% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_file.R
\name{sprinter}
\alias{sprinter}
\title{Creates a sprintf statement}
\usage{
sprinter(d, tvar, spvec, vnames)
}
\arguments{
\item{d}{input dataframe}

\item{tvar}{variable defining iterations}

\item{vnames}{variable names in d (excluding tvar) to print}

\item{sprintf}{format string}
}
\description{
Creates a sprintf statement
}
\keyword{internal}

