% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_soil_utilities.R
\name{srgf}
\alias{srgf}
\title{DSSAT45 root growth function}
\usage{
srgf(depth)
}
\arguments{
\item{depth}{Horizon depth}
}
\value{
Root distributions for DSSAT soil profiles
}
\description{
DSSAT45 root growth function
}
\keyword{internal}

