% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_utilities.R
\name{t_tab}
\alias{t_tab}
\title{Creates combination CSM X file treatment table}
\usage{
t_tab(tvars, topts, ttab)
}
\arguments{
\item{tvars}{Vector of variable names in treatments, e.g. c("INGENO, "PDATE")}

\item{topts}{Vector of corresponding factor level names, e.g. c("CU", "MP")}

\item{ttab}{Table of treatment permutations}
}
\description{
Creates combination CSM X file treatment table
}
\keyword{internal}
