% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{theta_s}
\alias{theta_s}
\title{Calculates saturated moisture content, requires function field_cap}
\usage{
theta_s(sand, clay, soc)
}
\arguments{
\item{sand}{Fraction of sand}

\item{clay}{Fraction of clay}

\item{soc}{soil organic matter percent}
}
\value{
Field capacity at 33 kpa
}
\description{
Calculates saturated moisture content, requires function field_cap
}
\keyword{internal}
