% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{theta_sdf}
\alias{theta_sdf}
\title{Calculates saturated water content, accounting for compaction}
\usage{
theta_sdf(sand, clay, soc, DF)
}
\arguments{
\item{sand}{Fraction of sand}

\item{clay}{Fraction of clay}

\item{soc}{Soil organic matter percent}

\item{DF}{Density factor between 0.9 and 1.3, no effect if set to 1}
}
\description{
Calculates saturated water content, accounting for compaction
}
\keyword{internal}
