% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weathdat}
\alias{weathdat}
\title{Weather data time series}
\format{A 3 element list consisting of two 2-element vectors and a data.table
with 11684 rows and 6 variables:
\itemize{
  \item dates: start and end dates of the weather data.
  \item xyz: latitude, longitude, and elevation of the location.
  \item dat: weather dataset
  \itemize{
    \item prec: precipitation (mm)
    \item tmax: maximum temperature (degress Celsius)
    \item tmin: minimum temperature (degress Celsius)
    \item sw: shortwave radiation (MJ/m^2/day) 
    \item wind: windspeed (km/day)
    \item pres: air pressure (kPa)
    \item sh: specific humidity (kg/kg)
  }
}}
\description{
A 30 year weather dataset for one location in southern Zambia
}
\examples{
data(weathdat)
}
\references{
Chaney, N.W., Sheffield, J., Villarini, G. & Wood, E.F. (2014)
Spatial Analysis of Trends in Climatic Extremes with a High Resolution 
Gridded Daily Meteorological Data Set over Sub-Saharan Africa. Journal of 
Climate, 27, 5815-5835.

Estes, L.D., Chaney, N.W., Herrera-Estrada, J., Sheffield, J., Caylor, K.K. 
& Wood, E.F. (2014) Changing water availability during the African 
maize-growing season, 1979–2010. Environmental Research Letters, 9, 075005.
}
\keyword{data}
