% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{Creates weather data file for crop model}
\usage{
weather(
  xy,
  elev,
  srad,
  tmax,
  tmin,
  prec,
  wind = NULL,
  ea = NULL,
  rh = NULL,
  tmu = NULL,
  sh = NULL,
  pres = NULL,
  tdew = NULL,
  sdate,
  edate,
  name,
  outdir,
  model = "DSSAT"
)
}
\arguments{
\item{xy}{vector giving longitude then latitude (degrees)}

\item{srad}{Solar radiation (mj/m^2/day)}

\item{tmax}{Maximum temperature (deg C)}

\item{tmin}{Minimum temperature (deg C)}

\item{prec}{Daily rainfall (mm)}

\item{ea}{Vapor pressure (mb, optional)}

\item{rh}{Relative humidity (\%, optional)}

\item{tmu}{Average temperature (deg C, optional)}

\item{sh}{Specific humidity (kg/kg, optional)}

\item{pres}{Surface pressure (mb, optional)}

\item{tdew}{Dewpoint temperature (deg C, optional)}

\item{sdate}{Start date for range in WTH file}

\item{edate}{End date for range in WTH file}

\item{name}{Name prefix for output weather data file}

\item{outdir}{Output directory for weather file}

\item{model}{Crop model type, default = "DSSAT"}
}
\value{
An output crop model weather file for a single location
}
\description{
Creates certain key inputs for crop model weather data files and
write them to appropriate format
}
\details{
This is currently only works for the DSSAT model, but is intended 
to be expanded to other models (e.g. APSIM). Note that this function assumes 
that daily data given to it will be the same date as UTC. This might be 
problematic in Australasia.
}
\examples{
sdate <- "19790701" 
edate <- "20100630"
xy <- c("y" = -16.2783358, "x" = 27.4738917) 
load("data/sp-clim-dat.rda")
elev <- 1120
srad <- swv
name <- "ZASP"
outdir <- "external/ext_data"
weather(xy, elev, srad, tmax, tmin, prec, wind = wind, sh = sh, 
        pres = pres, sdate = sdate, edate = edate, name = name, 
        outdir = outdir)
}
