% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weathergrid}
\alias{weathergrid}
\title{Weather data for gridding demonstration}
\format{A 22 element list consisting of 1 SpatialPolygonsDataFrame, 
a raster (20 cells), and 20 data.tables each containing 2191 rows and 4 
variables:
\itemize{
  \item namwala: SpatialPolygonsDataFrame showing district shape (EPSG 4326)
  \item wthgrid: RasterLayer subset from 0.25 degree grid, providing cell IDs
  \item C1960-1964, C2017-2021, C2074-2078, C2131-2135: data.tables of daily
  weather containing
  \itemize{
    \item sw: shortwave radiation (MJ/m^2/day) 
    \item tmax: maximum temperature (degress Celsius)
    \item tmin: minimum temperature (degress Celsius)
    \item prec: precipitation (mm)
  }
}}
\usage{
weathergrid
}
\description{
A dataset containing gridded daily weather data (2005-2010) over Namwala 
District in Zambia, as well as a shapefile and reference grid containing cell
IDs.  
The variables are as follows:
}
\references{
Chaney, N.W., Sheffield, J., Villarini, G. & Wood, E.F. (2014)
Spatial Analysis of Trends in Climatic Extremes with a High Resolution 
Gridded Daily Meteorological Data Set over Sub-Saharan Africa. Journal of 
Climate, 27, 5815-5835.

Estes, L.D., Chaney, N.W., Herrera-Estrada, J., Sheffield, J., Caylor, K.K. 
& Wood, E.F. (2014) Changing water availability during the African 
maize-growing season, 1979–2010. Environmental Research Letters, 9, 075005.
}
\keyword{datasets}

