% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedotransfer.R
\name{wilt_point}
\alias{wilt_point}
\title{Calculate wilting point}
\usage{
wilt_point(sand, clay, soc)
}
\arguments{
\item{sand}{Fraction of sand}

\item{clay}{Fraction of clay}

\item{soc}{Soil organic matter percent}
}
\value{
Wilting point at 1500 kpa
}
\description{
Calculate wilting point
}
\keyword{internal}
