% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wth.R
\name{write_wth}
\alias{write_wth}
\title{Write out DSSAT weather file}
\usage{
write_wth(
  dates,
  xy,
  elev,
  srad,
  tmax,
  tmin,
  prec,
  tdew = NULL,
  wind = NULL,
  outdir,
  name
)
}
\arguments{
\item{dates}{Vector of POSIX dates passed from weather function}

\item{xy}{vector giving longitude then latitude (degrees)}

\item{elev}{elevation of weather site/station (m)}

\item{srad}{Solar radiation (mj/m^2/day)}

\item{tmax}{Maximum temperature (deg C)}

\item{tmin}{Minimum temperature (deg C)}

\item{prec}{Daily rainfall (mm)}

\item{wind}{Wind speed (m/s, optional)}

\item{outdir}{Output directory name}

\item{name}{Name for output wth file}

\item{tmu}{Vector of mean temperatures (deg C)}

\item{tdewe}{Dew point temperate (deg C, optional)}
}
\value{
A printed WTH file in the specifed location
}
\description{
Write out a weather file for DSSAT
}
\details{
For now this function writes out -99 values for DEWP and WIND if 
missing. This could be made more efficient if these variables could be 
dropped.
}
\keyword{DSSAT}
