% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_utilities.R
\name{wthname}
\alias{wthname}
\title{Create unique four character DSSAT WTH file identifier}
\usage{
wthname(n, i)
}
\arguments{
\item{n}{Number of names needed}

\item{i}{index number of first prefix in subset (e.g. 1 corresponds to AAAA)}
}
\description{
Create unique four character DSSAT WTH file identifier
}
\keyword{internal}
