% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_file.R
\name{x_file}
\alias{x_file}
\title{Creates a DSSAT xfile}
\usage{
x_file(xtab, outdir, z, xtype)
}
\arguments{
\item{xtab}{CSM parameter input table}

\item{outdir}{Output directory name}

\item{z}{Two digit code for end of X file, typically 01}

\item{xtype}{CSM xfile type (corresponding to particular crop)}
}
\value{
A CSM experimental file, written to the location of choice
}
\description{
Creates a DSSAT xfile
}
\details{
This allows multiple treatments to be tested, but only for a single
field. As such, it is more appropriate, for the time being, to run the model 
for individual crops, rather than in spatial mode (.GSX) with multiple crops.
}
\note{
As presently written (4/3/15), the simulation controls will not work
for multiple methods, as a you need a complete general and automatic section
replicated for each treatment
}
