% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_tab.R
\name{x_tab}
\alias{x_tab}
\title{Creates input table of DSSAT CSM model parameters}
\usage{
x_tab(fields)
}
\arguments{
\item{fields}{data.frame/data.table with minimum variable set (see Details).}
}
\value{
Table of management variables joined to master table defining fields
}
\description{
Creates input table of variables to be passed to CSM xfile
writing function, joining them on to master input table
}
\details{
Function sets up a table containing all variables necessary to 
create a DSSAT X file. The minimum variables needed are: 
\itemize{
  \item ID_FIELD: field id (created by \code{\link{fid}})
  \item WTH: weather station id
  \item prof: soil profile id
  \item PDATE: Planting date (in YYDOY).
  \item XNAME: 4-digit X file identifier (created by \code{\link{xname}})
  \item SLB1-10: 10 columns for soil depths 
  \item SLLL1-10: 10 columns for wilting point values
  \item SDUL1-10: 10 columns for drained upper limit values
}
The last three sets of variables usually only have about 4-5 values, the rest
are NA. 10 columns are provided for the rare soil profile that has this many 
layers. These values are obtained by running \code{\link{read_sol_hor}} or 
\code{\link{read_sol_prof}} functions. Default values are provided for any 
variable beyond those required that are not specified. Note that the default
crop is maize, so if you don't specify a crop type in inputs that is what the
model will run. Additional variables specified in \emph{fields} will 
over-ride default values. Variable names must be exact or function will fail.
Other default settings worth knowing about: 
\itemize{
  \item SDATE, ICDAT, ODATE, RDATE are all assumed to equal PDATE, unless 
  otherwise specified.
  \item PFRST, PLAST For automatic planting, the planting window defaults to 
  opening 14 days before PDATE and closes 21 days afterwards. If variable
  planting dates is one of the treatments, this option should not be used. 
  For this reason, automatic planting is turned off by default. 
  \item FDATE (fertilization date) is assumed to be on day of planting. 
}
}
