% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_utilities.R
\name{xname}
\alias{xname}
\title{X file unique identifiers}
\usage{
xname(wth, n)
}
\arguments{
\item{wth}{vector of weather file name (e.g. ZASP7932.WTH)}

\item{n}{length of vector}
}
\description{
X file unique identifiers
}
\details{
Designed for data.table with use of BY statements
}
\keyword{internal}

